+++ rules/aip0134/response_message_name_test.go
@@ -68,3 +68,61 @@ func TestResponseMessageName(t *testing.T) {
 		})
 	}
 }
+
+func TestResponseMessageName_FullyQualified(t *testing.T) {
+	for _, test := range []struct {
+		name              string
+		TypePkg           string
+		ServicePkg        string
+		TypeName          string
+		ResponseTypeValue string
+		problems          testutils.Problems
+	}{
+		{
+			name:       "ValidLocalImport",
+			TypePkg:    "library",
+			ServicePkg: "library",
+			TypeName:   "Book",
+			problems:   nil,
+		},
+		{
+			name:       "ValidXPkgImport",
+			TypePkg:    "other",
+			ServicePkg: "library",
+			TypeName:   "Book",
+			problems:   nil,
+		},
+	} {
+		t.Run(test.name, func(t *testing.T) {
+			files := testutils.ParseProto3Tmpls(t, map[string]string{
+				"type.proto": `
+			package {{.TypePkg}};
+	
+			message {{.TypeName}} {}
+			`,
+				"service.proto": `
+			package {{.ServicePkg}};
+	
+			import "google/longrunning/operations.proto";
+			import "type.proto";
+	
+			service Foo {
+				rpc Update{{.TypeName}} (Update{{.TypeName}}Request) returns (google.longrunning.Operation) {
+					option (google.longrunning.operation_info) = {
+					response_type: "{{.TypePkg}}.{{.TypeName}}"
+					metadata_type: "Update{{.TypeName}}Metadata"
+					};
+				}
+			}
+			message Update{{.TypeName}}Request {}
+			message Update{{.TypeName}}Metadata {}
+			`,
+			}, test)
+			file := files["service.proto"]
+			got := responseMessageName.Lint(file)
+			if diff := test.problems.SetDescriptor(file.GetServices()[0].GetMethods()[0]).Diff(got); diff != "" {
+				t.Error(diff)
+			}
+		})
+	}
+}